Fix for Toast Crashes (#1)
I'll call out the specific issue below. This PR does a bunch - it updates the libraries to look like "ToastNotifications-Axion", uses a new version number, and updates the linkages/assembly references between Toast Notifications and the Toast Notifications Messages class.

The gist of the problem was that the anchor element that the toast messages is relative to will sometimes not be bound to a presentation object. I don't fully know what that means. But when it does, an unexpected exception is thrown and our app crashes.

This was reported in a few bugs:

NAV-4893
NAV-4596

I have submitted part of this code change to the actual Toast Notifications library people. But, in the meantime, this would let us get a patched version in our code.

The hope is that if accepted, we'll just switch to their library as soon as this is sorted out.